requirements.txt
# Install with: `pip install -r requirements.txt`

django
djangorestframework
MySQL-python
